Refactor retry into a separate module
